Make REST Exp boni floats
While none of the API keys currently violate this field's type, this behaviour is more consistent with the actual payloads returned in the event stream.

+++ auraxium/models/_metagame.py
@@ -23,4 +23,4 @@ class MetagameEventData(RESTPayload):
     name: LocaleData
     description: LocaleData
     type: int
-    experience_bonus: Optional[int] = None
+    experience_bonus: Optional[float] = None
